refactor(mobile): shared profile architecture with role context and a11y
