+++ apps/mobile/src/app/admin/common/profile.tsx
@@ -1,0 +1,5 @@
+import ProfileScreen from "../../../components/Profile/ProfileView";
+
+export default function AdminProfile() {
+  return <ProfileScreen userType="ADMIN" />;
+}
