+++ apps/mobile/src/app/entrepreneur/common/profile.tsx
@@ -1,4 +1,4 @@
-import ProfileScreen from "../../../components/ProfileScreen";
+import ProfileScreen from "../../../components/Profile/ProfileView";
 
 export default function EntrepreneurProfile() {
   return <ProfileScreen userType="ENTREPRENEUR" />;
