+++ apps/mobile/src/app/tourist/common/profile.tsx
@@ -1,4 +1,4 @@
-import ProfileScreen from "../../../components/ProfileScreen";
+import ProfileScreen from "../../../components/Profile/ProfileView";
 
 export default function TouristProfile() {
   return <ProfileScreen userType="TOURIST" />;
